+++ CHANGELOG.md
@@ -12,11 +12,19 @@ and this project adheres to [Semantic Versioning][].
 
 ### Added
 
--   Pushed `get_extent` functionality upstream to `spatialdata` (#162)
+-   Multiscale image handling: user can specify a scale, else the best scale is selected automatically given the figure size and dpi (#164)
+-   Large images are automatically rasterized to speed up performance (#164)
+-   Added better error message for mismatch in cs and ax number (#185)
 
 ### Fixed
 
 -   Now dropping index when plotting shapes after spatial query (#177)
+
+## [0.0.6] - 2023-11-06
+
+### Added
+
+-   Pushed `get_extent` functionality upstream to `spatialdata` (#162)
 
 ## [0.0.5] - 2023-10-02